moved function utilities to their own module
